pyproject.toml
[project]
name = "docflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.7",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
]
